Multithreading integration
Added multithreading so that the textures for the quads displaying the images of galaxies would load
simultaneously as the user interacts with the application. Created a system that detects when a user clicks on any quad. This will allow to extract information about a specific galaxy and display it to the user. In addition added a crosshair which allows for easier navigation in the 'In Depth Universe Explorer' scene.